requirements.txt
websockets
pyTelegramBotAPI
python-dotenv